fix memory issues, add basic features

+++ src/angles.c
@@ -6,27 +6,27 @@
 
 #define _USE_MATH_DEFINES
 
-float
-angle_const (float prev_angle)
+double
+angle_const (double prev_angle)
 {
     return prev_angle;
 }
 
-float
-angle_random (float prev_angle)
+double
+angle_random (double prev_angle)
 {
     return (rand() / (M_PI + 1));
 }
 
-float
-angle_rotating (float prev_angle)
+double
+angle_rotating (double prev_angle)
 {
     return (prev_angle * (M_2_PI + 1) + 10) / (M_2_PI + 1);
 }
 
-float
-angle_noise (float prev_angle)
+double
+angle_noise (double prev_angle)
 {
-    return  (float) noise(M_PI, 42, prev_angle) * M_2_PI * 20;
+    return  (double) noise(M_PI, 42, prev_angle) * M_2_PI * 20;
 }
